Making some final touches
There are some things to fine tune, but the project is mostly done.

Here are the outstanding items:
Spacing between the labels and the checkboxes and radio buttons.
Align the values in name and e-mail - make them sit to the right (or right-justified)
The border between the "user_info" and "type" fieldsets  (between the the user information and newsletter type fieldsets.)
Fix the Zip code input field so that it's only 50% of the width at the others above the 768px breakpoint.
The spin selector still appears for the zip code in Firefox.

Submitting to the Treehouse slack group to seek feedback.